+++ workspaces/jfrog-artifactory/plugins/jfrog-artifactory/src/translations/fr.ts
@@ -13,22 +13,28 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
+
 import { createTranslationMessages } from '@backstage/core-plugin-api/alpha';
 import { jfrogArtifactoryTranslationRef } from './ref';
 
+/**
+ * fr translation for plugin.jfrog-artifactory.
+ * @public
+ */
 const jfrogArtifactoryTranslationFr = createTranslationMessages({
   ref: jfrogArtifactoryTranslationRef,
   messages: {
-    'page.title': 'Dépôt JFrog Artifactory : {{image}}',
-    'table.searchPlaceholder': 'Filtrer',
+    'page.title': 'JFrog Artifactory repository: {{image}}',
+    'table.searchPlaceholder': 'Filtre',
     'table.labelRowsSelect': 'Lignes',
     'table.columns.version': 'Version',
     'table.columns.repositories': 'Dépôts',
     'table.columns.manifest': 'Manifeste',
     'table.columns.modified': 'Modifié',
     'table.columns.size': 'Taille',
-    'table.emptyContent.message': "Aucune donnée n'a encore été ajoutée,",
-    'table.emptyContent.learnMore': 'apprenez à ajouter des données.',
+    'table.emptyContent.message':
+      'Aucune donnée n’a été ajouté pour l’instant,',
+    'table.emptyContent.learnMore': 'apprenez comment ajouter des données.',
     'manifest.sha256': 'sha256',
   },
 });