+++ workspaces/acr/plugins/acr/CHANGELOG.md
@@ -1,4 +1,10 @@
-### Dependencies
+# @backstage-community/plugin-acr
+
+## 1.21.1
+
+### Patch Changes
+
+- c074f91: Updated fr, it and ja translation files.
 
 ## 1.21.0